+++ src/main/java/org/braekpo1nt/mctmanager/games/event/states/PlayingColossalCombatState.java
@@ -5,6 +5,7 @@ import net.kyori.adventure.text.format.NamedTextColor;
 import net.kyori.adventure.text.format.TextDecoration;
 import net.kyori.adventure.title.Title;
 import org.braekpo1nt.mctmanager.games.event.EventManager;
+import org.braekpo1nt.mctmanager.games.event.states.delay.ToPodiumDelayState;
 import org.braekpo1nt.mctmanager.ui.UIUtils;
 import org.bukkit.Bukkit;
 import org.jetbrains.annotations.Nullable;
