+++ src/main/java/org/braekpo1nt/mctmanager/games/event/states/delay/ColossalCombatDelay.java
@@ -1,4 +1,4 @@
-package org.braekpo1nt.mctmanager.games.event.states;
+package org.braekpo1nt.mctmanager.games.event.states.delay;
 
 import net.kyori.adventure.text.Component;
 import net.kyori.adventure.text.event.ClickEvent;
@@ -6,6 +6,7 @@ import net.kyori.adventure.text.format.NamedTextColor;
 import net.kyori.adventure.text.format.TextDecoration;
 import org.braekpo1nt.mctmanager.games.GameManager;
 import org.braekpo1nt.mctmanager.games.event.EventManager;
+import org.braekpo1nt.mctmanager.games.event.states.PlayingColossalCombatState;
 import org.braekpo1nt.mctmanager.games.utils.GameManagerUtils;
 import org.braekpo1nt.mctmanager.ui.timer.Timer;
 import org.bukkit.entity.Player;
